Split services layer to use case and transport layer
