+++ internal/services/middleware.go
@@ -1,0 +1,33 @@
+package services
+
+import (
+	"context"
+	"github.com/dgrijalva/jwt-go"
+	"log"
+	"net/http"
+)
+
+func (s *ToDoService) validToken(req *http.Request) (*http.Request, bool) {
+	token := req.Header.Get("Authorization")
+
+	claims := make(jwt.MapClaims)
+	t, err := jwt.ParseWithClaims(token, claims, func(*jwt.Token) (interface{}, error) {
+		return []byte(s.JWTKey), nil
+	})
+	if err != nil {
+		log.Println(err)
+		return req, false
+	}
+
+	if !t.Valid {
+		return req, false
+	}
+
+	id, ok := claims["user_id"].(string)
+	if !ok {
+		return req, false
+	}
+
+	req = req.WithContext(context.WithValue(req.Context(), userAuthKey(0), id))
+	return req, true
+}
