+++ internal/storages/entities.go
@@ -1,26 +1,27 @@
 package storages
 
 import (
+	"database/sql"
 	"time"
 )
 
 // Task reflects tasks in DB
 type Task struct {
-	ID        string    `json:"id"`
-	Content   string    `json:"content"`
-	UserID    string    `json:"user_id"`
-	CreatedAt time.Time `json:"created_at"`
-	UpdatedAt time.Time `json:"updated_at"`
-	DeletedAt time.Time `json:"deleted_at"`
+	ID        string       `json:"id"`
+	Content   string       `json:"content"`
+	UserID    string       `json:"user_id"`
+	CreatedAt time.Time    `json:"created_at"`
+	UpdatedAt time.Time    `json:"updated_at"`
+	DeletedAt sql.NullTime `json:"deleted_at"`
 }
 
 // User reflects users data from DB
 type User struct {
-	ID        string    `json:"id"`
-	Password  string    `json:"password"`
-	CreatedAt time.Time `json:"created_at"`
-	UpdatedAt time.Time `json:"updated_at"`
-	DeletedAt time.Time `json:"deleted_at"`
+	ID        string       `json:"id"`
+	Password  string       `json:"password"`
+	CreatedAt time.Time    `json:"created_at"`
+	UpdatedAt time.Time    `json:"updated_at"`
+	DeletedAt sql.NullTime `json:"deleted_at"`
 }
 
 type Login struct {
@@ -35,6 +36,11 @@ type CreateTask struct {
 func GetTask() Task {
 	var task Task
 	return task
+}
+
+func GetTasks() []Task {
+	var tasks []Task
+	return tasks
 }
 
 func GetUser() User {
